+++ peek_plugin_graphdb/_private/storage/ItemKeyIndexEncodedChunk.py
@@ -1,6 +1,6 @@
 import logging
 
-from sqlalchemy import Column, LargeBinary
+from sqlalchemy import Column, LargeBinary, BigInteger
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer, String
 from sqlalchemy.orm import relationship
@@ -23,7 +23,7 @@ class ItemKeyIndexEncodedChunk(Tuple, DeclarativeBase,
     __tablename__ = 'ItemKeyIndexEncodedChunk'
     __tupleType__ = graphDbTuplePrefix + 'ItemKeyIndexEncodedChunkTable'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
 
     modelSetId = Column(Integer,
                         ForeignKey('GraphDbModelSet.id', ondelete='CASCADE'),
